Split views into separate files.
