+++ django_app/redbox_app/redbox_core/views/__init__.py
@@ -1,0 +1,26 @@
+from redbox_app.redbox_core.views.chat_views import ChatsTitleView, ChatsView, RatingsView, post_message
+from redbox_app.redbox_core.views.citation_views import CitationsView
+from redbox_app.redbox_core.views.demographics_views import CheckDemographicsView, DemographicsView
+from redbox_app.redbox_core.views.document_views import (
+    UploadView,
+    documents_view,
+    file_status_api_view,
+    remove_doc_view,
+)
+from redbox_app.redbox_core.views.misc_views import health, homepage_view
+
+__all__ = [
+    "ChatsTitleView",
+    "ChatsView",
+    "CheckDemographicsView",
+    "CitationsView",
+    "DemographicsView",
+    "RatingsView",
+    "UploadView",
+    "documents_view",
+    "file_status_api_view",
+    "health",
+    "homepage_view",
+    "post_message",
+    "remove_doc_view",
+]
