+++ django_app/redbox_app/redbox_core/views/misc_views.py
@@ -1,0 +1,22 @@
+import logging
+
+from django.http import HttpRequest, HttpResponse
+from django.shortcuts import render
+from django.views.decorators.http import require_http_methods
+
+logger = logging.getLogger(__name__)
+
+
+@require_http_methods(["GET"])
+def homepage_view(request):
+    return render(
+        request,
+        template_name="homepage.html",
+        context={"request": request},
+    )
+
+
+@require_http_methods(["GET"])
+def health(_request: HttpRequest) -> HttpResponse:
+    """this required by ECS Fargate"""
+    return HttpResponse(status=200)
